Create README.md
I call it weather app. Which gives you a search bar to enter the name of a city. On hit of search, it shows a list of the weather forecast for the next 5 days with data for every 3 hours.

The Android app uses ViewModel, LiveData, Room architecture components. Follows the Android guide to architecture. It used Retrofit to do rest calls.